test(select): migrate states
