+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-disabled-placeholder-ios-ltr-Mobile-Chrome-linux.png
