+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-disabled-value-md-ltr-Mobile-Firefox-linux.png
