+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-placeholder-ios-ltr-Mobile-Safari-linux.png
