+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-placeholder-md-ltr-Mobile-Firefox-linux.png
