+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-value-ios-ltr-Mobile-Chrome-linux.png
