+++ core/src/components/select/test/states/select.e2e.ts-snapshots/select-value-md-ltr-Mobile-Safari-linux.png
